+++ packages/fake-data/.swcrc
