+++ packages/fake-data/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@tamanu/data-generation",
+  "name": "@tamanu/fake-data",
   "version": "2.26.0",
   "private": true,
   "description": "TODO",
