+++ packages/scripts/package.json
@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "@tamanu/database": "*",
-    "@tamanu/data-generation": "*",
+    "@tamanu/fake-data": "*",
     "commander": "^9.0.0",
     "config": "^3.3.9",
     "google-spreadsheet": "^2.0.7",
